Numeric.Data.Fraction

Description

Provides the Fraction type, a safer alternative to Ratio.

Since: 0.1

Type

data Fraction a where

Type for representing fractions. Designed to be similar to Ratio with the primary difference that it does not require the following invariants for its instances (e.g. Eq) to be sensible:

n / d is maximally reduced.

d > 0.

This has a number of consequences.

Fraction's Eq is based on an equivalence class, in contrast to Ratio, which compares the numerator and denominator directly:

Fractions are reduced first, e.g., 2 :%: 4 === 1 :%: 2.

Negative denominators are considered: 1 :%: 1 === -1 :%: -1.

The denominator is given more consideration:

abs operates on the numerator and the denominator.

signum is positive if both are negative.

Show x does not reduce x first. This is to make debugging easier.

Fraction Integer is a Field, and Fraction Natural is a Semiring.

Examples

Expand

>>> 2 %! 6 == 1 %! 3
True

>>> 1 %! 1 == -1 %! -1
True

>>> 1 %! 7 >= 1 %! -2
True

>>> -1 %! 7 >= 1 %! -2
True

>>> import Data.Text.Display (display)
>>> display $ 2 %! 6
"1 / 3"

Since: 0.1

Bundled Patterns

pattern (:%:) :: a -> a -> Fraction a infixr 5
Unidirectional pattern synonym for Fraction. This allows us to pattern match on a fraction term without exposing the unsafe internal details.

Since: 0.1

pattern (:%!) :: (HasCallStack, MEuclidean a, Normed a, Ord a, Semiring a, UpperBoundless a) => a -> a -> Fraction a infixr 5
Bidirectional pattern synonym for Fraction. Note that this is not safe in general, as construction with a zero denominator with throw an error.

WARNING: Partial

Since: 0.1

Creation

mkFraction :: (MEuclidean a, Normed a, Ord a, Semiring a, UpperBoundless a) => a -> a -> Either String (Fraction a)

Smart constructor for Fraction. Returns Nothing if the second parameter is 0. Reduces the fraction via reduce if possible.

Examples

Expand

>>> mkFraction 10 4
Right (UnsafeFraction 5 2)

>>> mkFraction 10 0
Left "Numeric.Data.Fraction: Fraction has zero denominator"

Since: 0.1

mkFractionTH :: (Lift a, MEuclidean a, Normed a, Ord a, Semiring a, UpperBoundless a) => a -> a -> Code Q (Fraction a)

Template haskell for creating a Fraction at compile-time.

Examples

Expand

>>> $$(mkFractionTH 7 2)
UnsafeFraction 7 2

Since: 0.1

(%%) :: (Lift a, MEuclidean a, Normed a, Ord a, Semiring a, UpperBoundless a) => a -> a -> Code Q (Fraction a) infixl 7

Infix version of mkFractionTH.

Examples

Expand

>>> $$(7 %% 2)
UnsafeFraction 7 2

Since: 0.1

unsafeFraction :: (HasCallStack, MEuclidean a, Normed a, Ord a, Semiring a, UpperBoundless a) => a -> a -> Fraction a

Throws an error when given a denominator of 0.

WARNING: Partial

Examples

Expand

>>> unsafeFraction 7 2
UnsafeFraction 7 2

Since: 0.1

(%!) :: (HasCallStack, MEuclidean a, Normed a, Ord a, Semiring a, UpperBoundless a) => a -> a -> Fraction a infixl 7

Infix version of unsafeFraction.

WARNING: Partial

Examples

Expand

>>> 7 %! 2
UnsafeFraction 7 2

Since: 0.1

Elimination

numerator :: Fraction a -> a

Since: 0.1

denominator :: Fraction a -> a

Since: 0.1

Functions

reduce :: (MEuclidean a, Normed a, Ord a, Semiring a, UpperBoundless a) => Fraction a -> Fraction a

Reduces a fraction:

Removes common factors.

Factors out negative denominators.

reduce (0 :%: _) --> 0 :%: 1.

Examples

Expand

>>> reduce (7 %! 2)
UnsafeFraction 7 2

>>> reduce (18 %! 10)
UnsafeFraction 9 5

>>> reduce (-5 %! -5)
UnsafeFraction 1 1

Since: 0.1

Optics

We provide a ReversedPrism' _MkFraction that allows for total elimination and partial construction, along with LabelOptic instances for "numerator" and "denominator".

Examples

Expand

>>> :set -XOverloadedLabels
>>> import Optics.Core (set, view)
>>> let x = 2 %! 7
>>> view #numerator x
2

>>> set #numerator 5 x
UnsafeFraction 5 7

>>> view #denominator x
7

_MkFraction :: (MEuclidean a, Normed a, Ord a, Semiring a, UpperBoundless a) => ReversedPrism' (Fraction a) (a, a)

ReversedPrism' that enables total elimination and partial construction.

Examples

Expand

>>> import Optics.Core (view)
>>> f = $$(2 %% 8)
>>> view _MkFraction f
(1,4)

>>> rmatching _MkFraction (0, 4)
Right (UnsafeFraction 0 1)

>>> rmatching _MkFraction (1, 0)
Left (1,0)

Since: 0.1

rmatching :: (Is (ReversedOptic k) An_AffineTraversal, ReversibleOptic k) => Optic k NoIx b a t s -> s -> Either t a

Reversed matching. Useful with smart-constructor optics.

Since: 0.1