Numeric.Data.Fraction.Internal

Description

Provides the Fraction type, a safer alternative to Ratio.

Since: 0.1

Type

data Fraction a

Type for representing fractions. Designed to be similar to Ratio with the primary difference that it does not require the following invariants for its instances (e.g. Eq) to be sensible:

n / d is maximally reduced.

d > 0.

This has a number of consequences.

Fraction's Eq is based on an equivalence class, in contrast to Ratio, which compares the numerator and denominator directly:

Fractions are reduced first, e.g., 2 :%: 4 === 1 :%: 2.

Negative denominators are considered: 1 :%: 1 === -1 :%: -1.

The denominator is given more consideration:

abs operates on the numerator and the denominator.

signum is positive if both are negative.

Show x does not reduce x first. This is to make debugging easier.

Fraction Integer is a Field, and Fraction Natural is a Semiring.

Examples

Expand

>>> 2 %! 6 == 1 %! 3
True

>>> 1 %! 1 == -1 %! -1
True

>>> 1 %! 7 >= 1 %! -2
True

>>> -1 %! 7 >= 1 %! -2
True

>>> import Data.Text.Display (display)
>>> display $ 2 %! 6
"1 / 3"

Since: 0.1

Constructors

UnsafeFraction !a !a

Bundled Patterns

pattern (:%:) :: a -> a -> Fraction a infixr 5
Unidirectional pattern synonym for Fraction. This allows us to pattern match on a fraction term without exposing the unsafe internal details.

Since: 0.1

pattern (:%!) :: (HasCallStack, MEuclidean a, Normed a, Ord a, Semiring a, UpperBoundless a) => a -> a -> Fraction a infixr 5
Bidirectional pattern synonym for Fraction. Note that this is not safe in general, as construction with a zero denominator with throw an error.

WARNING: Partial

Since: 0.1

Creation

unsafeFraction :: (HasCallStack, MEuclidean a, Normed a, Ord a, Semiring a, UpperBoundless a) => a -> a -> Fraction a

Throws an error when given a denominator of 0.

WARNING: Partial

Examples

Expand

>>> unsafeFraction 7 2
UnsafeFraction 7 2

Since: 0.1

(%!) :: (HasCallStack, MEuclidean a, Normed a, Ord a, Semiring a, UpperBoundless a) => a -> a -> Fraction a infixl 7

Infix version of unsafeFraction.

WARNING: Partial

Examples

Expand

>>> 7 %! 2
UnsafeFraction 7 2

Since: 0.1

Elimination

numerator :: Fraction a -> a

Since: 0.1

denominator :: Fraction a -> a

Since: 0.1

Functions

reduce :: (MEuclidean a, Normed a, Ord a, Semiring a, UpperBoundless a) => Fraction a -> Fraction a

Reduces a fraction:

Removes common factors.

Factors out negative denominators.

reduce (0 :%: _) --> 0 :%: 1.

Examples

Expand

>>> reduce (7 %! 2)
UnsafeFraction 7 2

>>> reduce (18 %! 10)
UnsafeFraction 9 5

>>> reduce (-5 %! -5)
UnsafeFraction 1 1

Since: 0.1

Misc

errMsg :: String

Since: 0.1