Numeric.Data.Interval

Description

Provides types for enforcing minimum and maximum bounds.

Since: 0.1

Left/Right-Interval

Type

data LRInterval l r a where

Represents a closed interval that is bounded on both sides i.e. LRInterval @l @r x represents \( x \in [l, r] \).

Since: 0.1

Bundled Patterns

pattern MkLRInterval :: a -> LRInterval l r a
Unidirectional pattern synonym for LRInterval. This allows us to pattern match on an interval term without exposing the unsafe internal details.

Since: 0.1

Creation

mkLRInterval :: forall l r a. (KnownNat l, KnownNat r, Num a, Ord a) => a -> Maybe (LRInterval l r a)

Smart constructor for LRInterval. Returns Nothing if the given value is not within the bounds.

Examples

Expand

>>> mkLRInterval @10 @100 50
Just (UnsafeLRInterval {unLRInterval = 50})

>>> mkLRInterval @10 @100 5
Nothing

>>> mkLRInterval @10 @100 200
Nothing

Since: 0.1

mkLRIntervalTH :: forall l r a. (Integral a, KnownNat l, KnownNat r, Lift a, Show a) => a -> Code Q (LRInterval l r a)

Template haskell for creating an LRInterval at compile-time.

Examples

Expand

>>> $$(mkLRIntervalTH @0 @100 7)
UnsafeLRInterval {unLRInterval = 7}

Since: 0.1

unsafeLRInterval :: forall l r a. (HasCallStack, KnownNat l, KnownNat r, Num a, Ord a, Show a) => a -> LRInterval l r a

Variant of mkLRInterval that throws an error when given a value out of bounds.

WARNING: Partial

Examples

Expand

>>> unsafeLRInterval @0 @10 7
UnsafeLRInterval {unLRInterval = 7}

Since: 0.1

reallyUnsafeLRInterval :: a -> LRInterval l r a

This function is an alias for the unchecked constructor UnsafeLRInterval i.e. it allows us to construct a LRInterval without checking the invariant. This is intended only for when we absolutely know the invariant holds and a branch (i.e. unsafeLRInterval) is undesirable for performance reasons. Exercise extreme caution.

Since: 0.1

Elimination

unLRInterval :: LRInterval l r a -> a

Since: 0.1

Optics

_MkLRInterval :: (KnownNat l, KnownNat r, Num a, Ord a) => ReversedPrism' (LRInterval l r a) a

ReversedPrism' that enables total elimination and partial construction.

Examples

Expand

>>> import Optics.Core ((^.))
>>> x = $$(mkLRIntervalTH @1 @5 2)
>>> x ^. _MkLRInterval
2

>>> rmatching (_MkLRInterval @1 @5) 3
Right (UnsafeLRInterval {unLRInterval = 3})

>>> rmatching (_MkLRInterval @1 @5) 7
Left 7

Since: 0.1

rmatching :: (Is (ReversedOptic k) An_AffineTraversal, ReversibleOptic k) => Optic k NoIx b a t s -> s -> Either t a

Left-Interval

Type

data LInterval l a where

Represents a closed interval that is left-bounded i.e. LInterval @l x represents \( x \in [l, \infty) \).

Since: 0.1

Bundled Patterns

pattern MkLInterval :: a -> LInterval l a
Unidirectional pattern synonym for LInterval. This allows us to pattern match on an interval term without exposing the unsafe internal details.

Since: 0.1

Creation

mkLInterval :: forall l a. (KnownNat l, Num a, Ord a) => a -> Maybe (LInterval l a)

Smart constructor for LInterval. Returns Nothing if the given value is not >= the bound.

Examples

Expand

>>> mkLInterval @10 50
Just (UnsafeLInterval {unLInterval = 50})

>>> mkLInterval @10 5
Nothing

Since: 0.1

mkLIntervalTH :: forall l a. (Integral a, KnownNat l, Lift a, Show a) => a -> Code Q (LInterval l a)

Template haskell for creating a LInterval at compile-time.

Examples

Expand

>>> $$(mkLIntervalTH @0 7)
UnsafeLInterval {unLInterval = 7}

Since: 0.1

unsafeLInterval :: forall l a. (HasCallStack, KnownNat l, Num a, Ord a, Show a) => a -> LInterval l a

Variant of mkLInterval that throws an error when given a value out of bounds.

WARNING: Partial

Examples

Expand

>>> unsafeLInterval @0 7
UnsafeLInterval {unLInterval = 7}

Since: 0.1

reallyUnsafeLInterval :: a -> LInterval l a

This function is an alias for the unchecked constructor UnsafeLInterval i.e. it allows us to construct a LInterval without checking the invariant. This is intended only for when we absolutely know the invariant holds and a branch (i.e. unsafeLInterval) is undesirable for performance reasons. Exercise extreme caution.

Since: 0.1

Elimination

unLInterval :: LInterval l a -> a

Since: 0.1

Optics

_MkLInterval :: (KnownNat l, Num a, Ord a) => ReversedPrism' (LInterval l a) a

ReversedPrism' that enables total elimination and partial construction.

Examples

Expand

>>> import Optics.Core ((^.))
>>> x = $$(mkLIntervalTH @8 10)
>>> x ^. _MkLInterval
10

>>> rmatching (_MkLInterval @8) 10
Right (UnsafeLInterval {unLInterval = 10})

>>> rmatching (_MkLInterval @8) 5
Left 5

Since: 0.1

Right-Interval

Type

data RInterval r a where

Represents a closed interval that is right-bounded i.e. RInterval @r x represents \( x \in (-\infty, r] \).

Since: 0.1

Bundled Patterns

pattern MkRInterval :: a -> RInterval r a
Unidirectional pattern synonym for RInterval. This allows us to pattern match on an interval term without exposing the unsafe internal details.

Since: 0.1

Creation

mkRInterval :: forall r a. (KnownNat r, Num a, Ord a) => a -> Maybe (RInterval r a)

Smart constructor for RInterval. Returns Nothing if the given value is not <= the bound.

Examples

Expand

>>> mkRInterval @100 50
Just (UnsafeRInterval {unRInterval = 50})

>>> mkRInterval @0 5
Nothing

Since: 0.1

mkRIntervalTH :: forall r a. (Integral a, KnownNat r, Lift a, Show a) => a -> Code Q (RInterval r a)

Template haskell for creating an RInterval at compile-time.

Examples

Expand

>>> $$(mkRIntervalTH @100 7)
UnsafeRInterval {unRInterval = 7}

Since: 0.1

unsafeRInterval :: forall r a. (HasCallStack, KnownNat r, Num a, Ord a, Show a) => a -> RInterval r a

Variant of mkRInterval that throws an error when given a value out of bounds.

WARNING: Partial

Examples

Expand

>>> unsafeRInterval @10 7
UnsafeRInterval {unRInterval = 7}

Since: 0.1

reallyUnsafeRInterval :: a -> RInterval r a

This function is an alias for the unchecked constructor UnsafeRInterval i.e. it allows us to construct a RInterval without checking the invariant. This is intended only for when we absolutely know the invariant holds and a branch (i.e. unsafeRInterval) is undesirable for performance reasons. Exercise extreme caution.

Since: 0.1

Elimination

unRInterval :: RInterval r a -> a

Since: 0.1

Optics

_MkRInterval :: (KnownNat r, Num a, Ord a) => ReversedPrism' (RInterval r a) a

ReversedPrism' that enables total elimination and partial construction.

Examples

Expand

>>> import Optics.Core ((^.))
>>> x = $$(mkRIntervalTH @8 5)
>>> x ^. _MkRInterval
5

>>> rmatching (_MkRInterval @8) 5
Right (UnsafeRInterval {unRInterval = 5})

>>> rmatching (_MkRInterval @8) 10
Left 10

Since: 0.1