Numeric.Data.ModP.Internal

Description

Internal tools for modular arithmetic and primality testing. The main functions are isPrime and findInverse, though others are exported for testing.

Since: 0.1

Primality Testing

data MaybePrime

Result of running Miller-Rabin algorithm. At best we can determine if some n is definitely composite or "probably prime".

Since: 0.1

Constructors

Composite

ProbablyPrime

isPrime :: Integer -> MaybePrime

Tests primality via the Miller-Rabin algorithm with 100 trials. Returns Composite if the number is definitely composite, otherwise ProbablyPrime.

Examples

Expand

>>> isPrime 7
ProbablyPrime

>>> isPrime 22
Composite

>>> isPrime 373
ProbablyPrime

Since: 0.1

isPrimeTrials :: Word16 -> Integer -> MaybePrime

isPrime that takes in an additional Word16 parameter for the number of trials to run. The more trials, the more confident we can be in ProbablyPrime.

Examples

Expand

>>> isPrimeTrials 1 91
ProbablyPrime

>>> isPrimeTrials 2 91
Composite

Note: False positives can be found via:

-- search for "ProbablyPrime" after 1 trial in the composite sequence
-- for a given prime p
counter p = filter ((== ProbablyPrime) . snd) $
  fmap (x -> (x, isPrimeTrials 1 x)) [p + p, p + p + p ..]

Since: 0.1

millerRabin :: Modulus -> Word16 -> MaybePrime

Miller-Rabin algorithm. Takes in the \(n\) to be tested and the number of trials to perform. The higher the trials, the higher our confidence in ProbablyPrime.

Helper Types

For the following functions/types, a core concept is rewriting our \(n\) as

\[ n = 2^r d + 1, \]

where \(d\) is odd i.e. we have factored out 2 as much as possible. We use newtypes to track these numbers.

newtype Modulus

Represents a modulus. When testing for primality, this is the \(n\) in \(n = 2^{r} d + 1\).

Since: 0.1

Constructors

MkModulus Integer

newtype Pow

The \(r\) in \(n = 2^{r} d + 1\).

Since: 0.1

Constructors

MkPow Integer

newtype Mult

The \(d\) in \(n = 2^{r} d + 1\).

Since: 0.1

Constructors

MkMult Integer

newtype Rand

Randomly generated \(m \in [2, n - 2] \) for testing \(n\)'s primality.

Since: 0.1

Constructors

MkRand Integer

Helper Functions

isWitness :: Modulus -> Pow -> Rand -> MaybePrime

For \(n, r, x\) with \(n = 2^{r} d + 1\) and some \(x \equiv a^d \pmod n \), returns Composite if \(x\) is a witness to \(n\) being composite. Otherwise returns ProbablyPrime.

Examples

Expand

>>> let (pow, mult) = factor2 (12 - 1)
>>> let testVal = 3 ^ mult `mod` 12
>>> isWitness 12 pow testVal
Composite

>>> let (pow, mult) = factor2 (7 - 1)
>>> let testVal = 3 ^ mult `mod` 7
>>> isWitness 7 pow testVal
ProbablyPrime

Since: 0.1

sqProgression :: Modulus -> Integer -> [Integer]

For \(n, x\), returns the infinite progression

\[ x, x^2, x^4, x^8, \ldots \pmod n. \]

Examples

Expand

>>> take 5 $ sqProgression 7 3
[3,2,4,2,4]

Since: 0.1

factor2 :: Modulus -> (Pow, Mult)

Given \(n\), returns \((r, d)\) such that \(n = 2^r d\) with \(d\) odd i.e. \(2\) has been factored out.

Examples

Expand

>>> factor2 7
(MkPow 0,MkMult 7)

>>> factor2 8
(MkPow 3,MkMult 1)

>>> factor2 20
(MkPow 2,MkMult 5)

Since: 0.1

Multiplicative Inverses

data Bezout

@since 0.1t

Constructors

MkBezout

Fields

bzGcd :: !R

bzS :: !S

bzT :: !T

newtype R

Since: 0.1

Constructors

R' Integer

newtype S

Since: 0.1

Constructors

S' Integer

newtype T

Since: 0.1

Constructors

T' Integer

findInverse :: Integer -> Modulus -> Integer

For \(a, p\), finds the multiplicative inverse of \(a\) in \(\mathbb{Z}/p\mathbb{Z}\). That is, finds e such that

\[ ae \equiv 1 \pmod p. \]

Note: The returned \(e\) is only an inverse when \(a\) and \(p\) are coprime i.e. \((a,p) = 1\). Of course this is guaranteed when \(p\) is prime and \(0 < a < p \), but it otherwise not true in general.

Also, this function requires division, it is partial when the modulus is 0.

Since: 0.1

findBezout :: Integer -> Modulus -> Bezout

Since: 0.1