Numeric.Data.Positive.Algebra

Description

Provides the Positive type for enforcing a positive invariant. In contrast to Numeric.Data.Positive.Base, typeclass instances are provided in terms of algebra-simple classes, where possible.

Since: 0.1

Type

data Positive a where

Newtype wrapper that attaches a Positive invariant to some a. Positive is an ASemigroup and MGroup i.e. supports addition, multiplication, and division.

Since: 0.1

Bundled Patterns

pattern MkPositive :: a -> Positive a
Unidirectional pattern synonym for Positive. This allows us to pattern match on a positive term without exposing the unsafe internal details.

Since: 0.1

Creation

mkPositiveTH :: (AMonoid a, Lift a, Ord a, Show a) => a -> Code Q (Positive a)

Template haskell for creating a Positive at compile-time.

Examples

Expand

>>> $$(mkPositiveTH 1)
UnsafePositive 1

Since: 0.1

mkPositive :: (AMonoid a, Ord a) => a -> Maybe (Positive a)

Smart constructor for Positive. Returns Nothing if the second parameter is <= 0.

Examples

Expand

>>> mkPositive 7
Just (UnsafePositive 7)

>>> mkPositive 0
Nothing

Since: 0.1

unsafePositive :: (AMonoid a, HasCallStack, Ord a, Show a) => a -> Positive a

Throws an error when given a value <= 0.

WARNING: Partial

Examples

Expand

>>> unsafePositive 7
UnsafePositive 7

Since: 0.1

(+!) :: (AMonoid a, HasCallStack, Ord a, Show a) => a -> Positive a infixl 7

Postfix operator for unsafePositive.

WARNING: Partial

Examples

Expand

>>> (7 +!)
UnsafePositive 7

Since: 0.1

reallyUnsafePositive :: a -> Positive a

This function is an alias for the unchecked constructor UnsafePositive i.e. it allows us to construct a Positive without checking the invariant. This is intended only for when we absolutely know the invariant holds and a branch (i.e. unsafePositive) is undesirable for performance reasons. Exercise extreme caution.

Since: 0.1

Elimination

unPositive :: Positive a -> a

Since: 0.1

Functions

positiveToNonZero :: Positive a -> NonZero (Positive a)

Convenience function for adding a NonZero proof to our Positive.

Examples

Expand

>>> positiveToNonZero $ unsafePositive 3
UnsafeNonZero (UnsafePositive 3)

Since: 0.1

Optics

We provide a ReversedPrism' _MkPositive that allows for total elimination and partial construction, along with a LabelOptic Getter for #unPositive.

Examples

Expand

>>> :set -XOverloadedLabels
>>> import Optics.Core (view)
>>> let n = $$(mkPositiveTH 2)
>>> view #unPositive n
2

_MkPositive :: (AMonoid a, Ord a) => ReversedPrism' (Positive a) a

ReversedPrism' that enables total elimination and partial construction.

Examples

Expand

>>> import Optics.Core (view)
>>> pos = $$(mkPositiveTH 2)
>>> view _MkPositive pos
2

>>> rmatching _MkPositive 3
Right (UnsafePositive 3)

>>> rmatching _MkPositive 0
Left 0

Since: 0.1

rmatching :: (Is (ReversedOptic k) An_AffineTraversal, ReversibleOptic k) => Optic k NoIx b a t s -> s -> Either t a

Reversed matching. Useful with smart-constructor optics.

Since: 0.1