Numeric.Data.Positive.Internal

Description

Provides the Positive type for enforcing a positive invariant.

Since: 0.1

Type

newtype Positive a

Newtype wrapper that attaches a Positive invariant to some a. Positive is an ASemigroup and MGroup i.e. supports addition, multiplication, and division.

Since: 0.1

Constructors

UnsafePositive a

Bundled Patterns

pattern MkPositive :: a -> Positive a
Unidirectional pattern synonym for Positive. This allows us to pattern match on a positive term without exposing the unsafe internal details.

Since: 0.1

Creation

unsafePositive :: (HasCallStack, Num a, Ord a, Show a) => a -> Positive a

Throws an error when given a value <= 0.

WARNING: Partial

Examples

Expand

>>> unsafePositive 7
UnsafePositive 7

Since: 0.1

Misc

errMsg :: Show a => String -> a -> String

Since: 0.1