Numeric.Data.Interval

Description

Provides types for enforcing minimum and maximum bounds.

Since: 0.1

Types

data IntervalBound

Interval bound.

Since: 0.1

Constructors

Open Nat
Open bound.

Closed Nat
Closed bound.

None
No bound.

data Interval (l :: IntervalBound) (r :: IntervalBound) (a :: Type) where

Represents an interval. Can be (open|closed) bounded (left|right).

Examples

Expand

>>> import Data.Text.Display (display)
>>> import Data.Text qualified as T
>>> let x = unsafeInterval @(Open 10) @(Closed 100) 50
>>> putStrLn $ T.unpack $ display x
50 ∈ (10, 100]

>>> let y = unsafeInterval @None @None (-2)
>>> putStrLn $ T.unpack $ display y
-2 ∈ (-∞, ∞)

Since: 0.1

Bundled Patterns

pattern MkInterval :: a -> Interval l r a

Creation

mkInterval :: forall l r a. (Num a, Ord a, SingIntervalBound l, SingIntervalBound r) => a -> Maybe (Interval l r a)

Smart constructor for Interval. Returns Nothing if the given value is not within the bounds. Note that we do not check that the bounds fit within the type itself (e.g. consider Interval None (Closed 200) Int8).

Examples

Expand

>>> mkInterval @(Open 10) @(Closed 100) 50
Just (UnsafeInterval (Open 10) (Closed 100) 50)

>>> mkInterval @(Open 10) @(Closed 100) 100
Just (UnsafeInterval (Open 10) (Closed 100) 100)

>>> mkInterval @(Open 10) @(Closed 100) 10
Nothing

>>> mkInterval @(Open 10) @(Closed 100) 101
Nothing

Since: 0.1

mkIntervalTH :: forall l r a. (Lift a, Num a, Ord a, SingIntervalBound l, SingIntervalBound r, Show a) => a -> Code Q (Interval l r a)

Template haskell for creating an Interval at compile-time.

Examples

Expand

>>> $$(mkIntervalTH @None @(Closed 100) 7)
UnsafeInterval None (Closed 100) 7

Since: 0.1

unsafeInterval :: forall l r a. (HasCallStack, Num a, Ord a, SingIntervalBound l, SingIntervalBound r, Show a) => a -> Interval l r a

Variant of mkInterval that throws an error when given a value out of bounds.

WARNING: Partial

Examples

Expand

>>> unsafeInterval @(Open 10) @(Closed 100) 50
UnsafeInterval (Open 10) (Closed 100) 50

Since: 0.1

reallyUnsafeInterval :: a -> Interval l r a

This function is an alias for the unchecked constructor UnsafeInterval i.e. it allows us to construct a Interval without checking invariants. This is intended only for when we absolutely know the invariant holds and a branch (i.e. unsafeInterval) is undesirable for performance reasons. Exercise extreme caution.

Since: 0.1

Elimination

unInterval :: Interval l r a -> a

Since: 0.1

Optics

_MkInterval :: forall l r a. (Num a, Ord a, SingIntervalBound l, SingIntervalBound r) => ReversedPrism' (Interval l r a) a

ReversedPrism' that enables total elimination and partial construction.

Examples

Expand

>>> import Optics.Core ((^.))
>>> x = $$(mkIntervalTH @(Open 1) @(Open 5) 2)
>>> x ^. _MkInterval
2

>>> rmatching (_MkInterval @(Open 1) @(Open 5)) 3
Right (UnsafeInterval (Open 1) (Open 5) 3)

>>> rmatching (_MkInterval @(Open 1) @(Open 5)) 7
Left 7

Since: 0.1

rmatching :: (Is (ReversedOptic k) An_AffineTraversal, ReversibleOptic k) => Optic k NoIx b a t s -> s -> Either t a

Reversed matching. Useful with smart-constructor optics.

Since: 0.1