Numeric.Data.ModP

Description

Provides the ModP type for modular arithmetic.

Since: 0.1

Type

data ModP p a where

Newtype wrapper that represents \( \mathbb{Z}/p\mathbb{Z} \) for prime p. ModP is a Field i.e. supports addition, subtraction, multiplication, and division.

Since: 0.1

Bundled Patterns

pattern MkModP :: a -> ModP p a
Unidirectional pattern synonym for ModP. This allows us to pattern match on a modp term without exposing the unsafe internal details.

Since: 0.1

Creation

mkModP :: forall p a. (KnownNat p, UpperBoundless a) => a -> Maybe (ModP p a)

Constructor for ModP. Fails if p is not prime. This uses the Miller-Rabin primality test, which has complexity \(O(k \log^3 p)\), and we take \(k = 100\). See wikipedia for more details.

Examples

Expand

>>> mkModP @5 7
Just (MkModP 2 (mod 5))

>>> mkModP @10 7
Nothing

Since: 0.1

mkModPTH :: forall p a. (KnownNat p, Lift a, UpperBoundless a) => a -> Code Q (ModP p a)

Template haskell for creating a ModP at compile-time.

Examples

Expand

>>> $$(mkModPTH @11 7)
MkModP 7 (mod 11)

Since: 0.1

unsafeModP :: forall p a. (HasCallStack, KnownNat p, UpperBoundless a) => a -> ModP p a

Variant of mkModP that throws an error when given a non-prime.

WARNING: Partial

Examples

Expand

>>> unsafeModP @7 12
MkModP 5 (mod 7)

Since: 0.1

reallyUnsafeModP :: forall p a. (KnownNat p, UpperBoundless a) => a -> ModP p a

This function reduces the argument modulo p but does not check that p is prime. Note that the correct behavior of some functionality (e.g. division) is reliant on primality, so this is dangerous. This is intended only for when we absolutely know p is prime and the check is undesirable for performance reasons. Exercise extreme caution.

Since: 0.1

Elimination

unModP :: ModP p a -> a

Since: 0.1

Functions

invert :: forall p a. (KnownNat p, UpperBoundless a) => NonZero (ModP p a) -> ModP p a

Given non-zero \(d\), returns the inverse i.e. finds \(e\) s.t.

\[ de \equiv 1 \pmod p. \]

Examples

Expand

>>> invert $ unsafeAMonoidNonZero $ unsafeModP @7 5
MkModP 3 (mod 7)

>>> invert $ unsafeAMonoidNonZero $ unsafeModP @19 12
MkModP 8 (mod 19)

Since: 0.1

Optics

_MkModP :: (KnownNat p, UpperBoundless a) => ReversedPrism' (ModP p a) a

ReversedPrism' that enables total elimination and partial construction.

Examples

Expand

>>> import Optics.Core ((^.))
>>> n = $$(mkModPTH @7 9)
>>> n ^. _MkModP
2

>>> rmatching (_MkModP @7) 9
Right (MkModP 2 (mod 7))

>>> rmatching (_MkModP @6) 9
Left 9

Since: 0.1

rmatching :: (Is (ReversedOptic k) An_AffineTraversal, ReversibleOptic k) => Optic k NoIx b a t s -> s -> Either t a