Numeric.Data.NonZero

Description

Provides the NonZero type for enforcing a non-zero invariant.

Since: 0.1

Type

data NonZero a where

Smart-constructor for creating a "non-zero" a.

Since: 0.1

Bundled Patterns

pattern MkNonZero :: a -> NonZero a
Unidirectional pattern synonym for NonZero. This allows us to pattern match on a nonzero term without exposing the unsafe internal details.

Since: 0.1

Creation

mkNonZero :: (Eq a, Num a) => a -> Maybe (NonZero a)

Smart constructor for NonZero.

Examples

Expand

>>> mkNonZero 7
Just (UnsafeNonZero 7)

>>> mkNonZero 0
Nothing

Since: 0.1

mkNonZeroTH :: (Eq a, Lift a, Num a) => a -> Code Q (NonZero a)

Template-haskell version of mkNonZero for creating NonZero at compile-time.

Examples

Expand

>>> $$(mkNonZeroTH 7)
UnsafeNonZero 7

Since: 0.1

unsafeNonZero :: (Eq a, HasCallStack, Num a) => a -> NonZero a

Throws an error when given 0.

WARNING: Partial

Examples

Expand

>>> unsafeNonZero 7
UnsafeNonZero 7

Since: 0.1

reallyUnsafeNonZero :: a -> NonZero a

This function is an alias for the unchecked constructor UnsafeNonZero i.e. it allows us to construct a NonZero without checking the invariant. This is intended only for when we absolutely know the invariant holds and a branch (i.e. unsafeNonZero) is undesirable for performance reasons. Exercise extreme caution.

Since: 0.1

Elimination

unNonZero :: NonZero a -> a

Since: 0.1

Optics

We provide a ReversedPrism' _MkNonZero that allows for total elimination and partial construction, along with a LabelOptic Getter for #unNonZero.

Examples

Expand

>>> :set -XOverloadedLabels
>>> import Optics.Core (view)
>>> let n = $$(mkNonZeroTH 7)
>>> view #unNonZero n
7

_MkNonZero :: (Eq a, Num a) => ReversedPrism' (NonZero a) a

ReversedPrism' that enables total elimination and partial construction.

Examples

Expand

>>> import Optics.Core (view)
>>> nz = $$(mkNonZeroTH 7)
>>> view _MkNonZero nz
7

>>> rmatching _MkNonZero 3
Right (UnsafeNonZero 3)

>>> rmatching _MkNonZero 0
Left 0

Since: 0.1

rmatching :: (Is (ReversedOptic k) An_AffineTraversal, ReversibleOptic k) => Optic k NoIx b a t s -> s -> Either t a

Reversed matching. Useful with smart-constructor optics.

Since: 0.1